Create worker crate
Move worker functionalities from configuration to new worker crate. Move
ConcernKey to new worker crate. Change dependencies to reflect that.
Change exit value of dispatcher to reflect if it was due to an error or
if it exited normally.

Create background system thread to poll for worker state. Transfer
the worker's balance back to the owner in case of retirement, and
exit the process with an Ok code.